move image&registry func to new file
